[test] local version of analytics.js
Again, just seeing if this is helpful in any way, then will look into a
PR that will allow the User to set their own version of the GA Script if
necessary.

TO be fair, ga-lite did not see to work (in initial testing) which is
why we are trying this now.