[feat] customJavascript option for non-Google Hosted
Added an option for `customJavascript` that will default to the Google
Hosted JavaScript file at
`https://www.google-analytics.com/analytics.js`, however, will allow a
User to set a custom location if they so choose.

Note: This does not work with `ga-lite`, and only a custom hosted
analytics solution.

Please also Note: If you decide to use this option, you are now solely
responsible for keeping up with Google releases.